separated out gym training into four different levels

+++ Assets/Scripts/ChildPower.cs
@@ -6,14 +6,14 @@ public class ChildPower {
 	//Need power name, mother power, father power, point values for str, agility, will, and fortitude, need base power popularity value, etc.
 	public string powerName { get; set; } 
 	public string description { get; set; }
-	public double strength { get; set; } //0 - 100
-	public double agility { get; set; } //0 - 100
-	public double will { get; set; }  //0 - 100
-	public double fortitude { get; set; } //0 - 100
-	public double popularity { get; set; } //0 - 100
+	public decimal strength { get; set; } //0 - 100
+	public decimal agility { get; set; } //0 - 100
+	public decimal will { get; set; }  //0 - 100
+	public decimal fortitude { get; set; } //0 - 100
+	public decimal popularity { get; set; } //0 - 100
 
 	public ChildPower (ParentalPower fPower, ParentalPower mPower) {
-    double[] childAttributes = getGenerationalAttributes(fPower, mPower);
+    decimal[] childAttributes = getGenerationalAttributes(fPower, mPower);
     string fPowerName = fPower.PowerName;
     string mPowerName = mPower.PowerName;
     strength = childAttributes[0];
@@ -25,7 +25,7 @@ public class ChildPower {
     if (fPowerName == "SUPER STRENGTH" && mPowerName == "SUPER STRENGTH") {
       powerName = "DEVESTATING MIGHT";
       description = "Your muscles crackle with power. Even at such a young age your muscles bulge, barely able to contain the raw might stored within them. You've had very little run-ins with villains up to this point, and those you've faced buckled under your sheer power. Your enemies know best not to cross you.";
-      strength = this.strength * 1.4f;
+      strength = this.strength * 1.4m;
     }
     if ((fPowerName == "SUPER STRENGTH" && mPowerName == "TELEKINESIS") || (fPowerName == "TELEKINESIS" && mPowerName == "SUPER STRENGTH")) {
       powerName = "CONQUER";
@@ -38,12 +38,12 @@ public class ChildPower {
     }
     if ((fPowerName == "SUPER STRENGTH" && mPowerName == "SHAPESHIFT") || (fPowerName == "SHAPESHIFT" && mPowerName == "SUPER STRENGTH")) {
       powerName = "RAVAGE FORM";
-      fortitude = this.fortitude * 1.2f;
+      fortitude = this.fortitude * 1.2m;
       description = "Your muscles bulge and tear, your bones protrude and harden. At will you are able to become a being of pure power: a carnal thing with a hunger for decimation. You have a form that you only refer to as 'Ravager'. Others know it as 'Monster'. Whatever title you wear one thing is certain: the world has not seen destruction as grand as you've made.";
     }
     if ((fPowerName == "TELEKINESIS" && mPowerName == "TELEKINESIS")) {
       powerName = "PSION";
-      will = this.will * 1.4f;
+      will = this.will * 1.4m;
       description = "Your parents were strong psychics in their own right, but you are even more than they could ever achieve. At a young age you could bend cars like hair in your fingers, growing up you could cause the clouds to stir on a whim. Your eyes see both the present and the future, a skill that neither of your parents possessed. You are an immensely powerful Psion. The world will bend at your feet whether they wish to or not.";
     }
     if ((fPowerName == "TELEKINESIS" && mPowerName == "PYROMANCY") || (fPowerName =="PYROMANCY" && mPowerName == "TELEKINESIS")) {
@@ -57,7 +57,7 @@ public class ChildPower {
     if (fPowerName == "PYROMANCY" && mPowerName == "PYROMANCY") {
       powerName = "FIRESTORM";
       description = "You are the sun, powerful and awesome. You are the light, radient and beautiful. You are the firestorm, consuming everything as you wish. The world burns with your steps. The flames will purify all. There can be nothing left in the wake of your fire.";
-      fortitude = this.fortitude * 1.4f;
+      fortitude = this.fortitude * 1.4m;
     }
     if ((fPowerName == "PYROMANCY" && mPowerName == "SHAPESHIFT") || fPowerName == "SHAPESHIFT" && mPowerName == "PYROMANCY") {
       powerName = "FLAME ASPECT";
@@ -66,17 +66,17 @@ public class ChildPower {
     if (fPowerName == "SHAPESHIFT" && mPowerName == "SHAPESHIFT") {
       powerName = "A THOUSAND FACES";
       description = "Your mastery over forms allows you to become anything you've ever dreamed of. It only takes one look for you to imitate someone completely.";
-      agility = this.agility * 1.4f;
+      agility = this.agility * 1.4m;
     }
 	}
 
-  double[] getGenerationalAttributes (ParentalPower father, ParentalPower mother) {
-    double cStrength = (((father.strength + mother.strength) /2) * 0.25f);
-    double cAgility = (((father.agility + mother.agility) /2) * 0.25f);
-    double cWill = (((father.will + mother.will) /2) * 0.25f);
-    double cFortitude = (((father.fortitude + mother.fortitude) /2) * 0.25f);
+  decimal[] getGenerationalAttributes (ParentalPower father, ParentalPower mother) {
+    decimal cStrength = (((father.strength + mother.strength) /2) * 0.25m);
+    decimal cAgility = (((father.agility + mother.agility) /2) * 0.25m);
+    decimal cWill = (((father.will + mother.will) /2) * 0.25m);
+    decimal cFortitude = (((father.fortitude + mother.fortitude) /2) * 0.25m);
 
-    double[] attributesArray = new double[] {
+    decimal[] attributesArray = new decimal[] {
       cStrength, cAgility, cWill, cFortitude
     };
